requirements.txt
# Core dependencies
streamlit==1.39.0
fastapi>=0.115.0
uvicorn==0.24.0

# LLM & AI
# vllm==0.2.7
torch
sentence-transformers
transformers

# RAG & Vector DB
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.38
# pymilvus==2.3.4
qdrant-client==1.7.0

# Document processing  
pymupdf==1.23.8
pdfplumber==0.10.3
paddleocr==2.7.3
Pillow>=10.3.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.13.1

# Utils
pydantic>=2.11.7,<3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests>=2.32.2
beautifulsoup4==4.12.2
numpy
pandas==2.1.4
uuid==1.30

# Development
jupyter==1.0.0
ipykernel>=6.29.3,<7
pytest==7.4.3
black==23.11.0

# packaging>=24.2
pydantic-settings
chardet